+++ src/components/Accordion/accordionChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Accessibility", "Functionality"],
+    notes: ["Adds logic to close accordion when 'esc' key is pressed"],
+  },
+  {
     date: "2024-03-14",
     version: "3.0.0",
     type: "Update",